terminamos de adaptar pra arquitetura
