+++ agendando_um_jeito/src/contexts/authContext.jsx
@@ -1,8 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import { onAuthStateChanged, signOut } from 'firebase/auth';
 import { createUserWithEmailAndPassword } from "firebase/auth";
-import { auth, db } from "../services/firebase";
-import { doc, setDoc } from "firebase/firestore";
+import { auth, provider, db } from "../services/firebase";
+import { useNavigate } from 'react-router-dom';
+import { doc, setDoc, updateDoc} from "firebase/firestore";
+import { signInWithPopup, signInWithEmailAndPassword } from 'firebase/auth';
 
 export const AuthContext = React.createContext({});
 
@@ -11,6 +13,7 @@ export const AuthProvider = ({ children }) => {
     const [user, setUser] = useState(null);
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
+    const navigate = useNavigate()
 
     useEffect(() => {
         const unsubscribe = onAuthStateChanged(auth, (user) => {
@@ -46,7 +49,35 @@ export const AuthProvider = ({ children }) => {
         }
     }
 
-    const value = { signed, setSigned, user, setUser, logout, Signup, email, setEmail, password, setPassword };
+    const SignIn = async (event) => {
+        event.preventDefault();
+        try {
+            const userCredential = await signInWithEmailAndPassword(auth, email, password);
+            const user = userCredential.user;
+            console.log(user)
+            setEmail("");
+            setPassword("");
+            navigate('/Home')
+        } catch (error) {
+            console.log("error: ", error);
+        }
+      };
+
+    const SignInWithGoogle = async () => {
+        try {
+            const userCredential = await signInWithPopup(auth, provider)
+                const user = userCredential.user
+            const email = user.email;
+            const usersCollectionRef = doc(db, 'users', user.uid);
+            await updateDoc(usersCollectionRef, { email, googleAuth: true });
+            navigate('/Home')
+    
+        }catch (error) {
+            console.log('error: ', error);
+        }
+    }
+
+    const value = { signed, setSigned, user, setUser, logout, Signup, email, setEmail, password, setPassword, SignIn, SignInWithGoogle};
 
     return (
         <AuthContext.Provider value={value}>
